fixed failing test in gameservice + recovery of changes in other game files

+++ GameShop/src/main/java/com/mcgill/ecse321/GameShop/model/Game.java
@@ -65,7 +65,7 @@ public class Game {
   // CONSTRUCTOR
   // ------------------------
 
-  public Game(String aTitle, String aDescription, int aPrice, GameStatus aGameStatus, int aStockQuantity,
+  public Game(String aTitle, String aDescription, double aPrice, GameStatus aGameStatus, int aStockQuantity,
       String aPhotoUrl) {
     title = aTitle;
     description = aDescription;
@@ -116,7 +116,7 @@ public class Game {
     return wasSet;
   }
 
-  public boolean setPrice(int aPrice) {
+  public boolean setPrice(double aPrice) {
     boolean wasSet = false;
     price = aPrice;
     wasSet = true;
